update TelerikMessages resource files

+++ src/Fims.Client.Shared/Pages/Management/UserManagement.razor.cs
@@ -1,11 +1,13 @@
 ﻿using Fims.Data.Entities;
 using Fims.Data.Models.Identity;
+using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Identity;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Telerik.Blazor;
 using Telerik.Blazor.Components;
 
 namespace Fims.Client.Shared.Pages.Management
@@ -62,28 +64,34 @@ namespace Fims.Client.Shared.Pages.Management
 
         private async Task DeleteUser(GridCommandEventArgs args)
         {
-            var userinfo = (UserAuthInfoModel)args.Item;
+            bool confirmed = await Dialogs.ConfirmAsync("해당 계정은 영구 삭제 됩니다.", "계정 삭제");
 
-            var result = await this.AuthClientService.Delete(userinfo.UserName);
-
-            if (result.Succeeded)
+            if (confirmed)
             {
-                this.ShowErrors = false;
+                var userinfo = (UserAuthInfoModel)args.Item;
 
-                UserManagementNotificationComponent.Show(new NotificationModel
+                var result = await this.AuthClientService.Delete(userinfo.UserName);
+
+                if (result.Succeeded)
                 {
-                    Text = "사용자 계정삭제 성공",
-                    ThemeColor = "success",
-                    //CloseAfter = 3000
-                });
-            }
-            else
-            {
-                this.Errors = result.Errors;
-                this.ShowErrors = true;
+                    this.ShowErrors = false;
+
+                    UserManagementNotificationComponent.Show(new NotificationModel
+                    {
+                        Text = "사용자 계정삭제 성공",
+                        ThemeColor = "success",
+                        //CloseAfter = 3000
+                    });
+                }
+                else
+                {
+                    this.Errors = result.Errors;
+                    this.ShowErrors = true;
+                }
+
+                await LoadData();
             }
 
-            await LoadData();
         }
 
         private void ShowAddUserDialog(GridCommandEventArgs args)
@@ -100,6 +108,59 @@ namespace Fims.Client.Shared.Pages.Management
             await LoadData();
         }
 
+
+        [CascadingParameter]
+        public DialogFactory Dialogs { get; set; }
+
+        public string Title { get; set; }
+
+        public string Message { get; set; }
+
+        public async Task ActivateAlert()
+        {
+            if (string.IsNullOrWhiteSpace(Title))
+            {
+                await Dialogs.AlertAsync("Something went wrong!");
+            }
+            else
+            {
+                await Dialogs.AlertAsync("Something went wrong!", Title);
+            }
+
+            Message = "The user saw and closed an Alert dialog.";
+        }
+
+        public async Task ActivateConfirm()
+        {
+            bool confirmed;
+            if (string.IsNullOrWhiteSpace(Title))
+            {
+                confirmed = await Dialogs.ConfirmAsync("Are you sure?");
+            }
+            else
+            {
+                confirmed = await Dialogs.ConfirmAsync("Are you sure?", Title);
+            }
+
+            var confirmMessage = confirmed ? "confirmed" : "canceled";
+
+            Message = $"The user {confirmMessage} the dialog.";
+        }
+
+        public async Task ActivatePrompt()
+        {
+            string input;
+            if (string.IsNullOrWhiteSpace(Title))
+            {
+                input = await Dialogs.PromptAsync("Please, enter your answer.");
+            }
+            else
+            {
+                input = await Dialogs.PromptAsync("Please, enter your answer.", Title);
+            }
+
+            Message = $"The returned user input is \"{input}\".";
+        }
 
     }
 }
